split out worker implementation
